Move annotation CDF connection
