+++ react-components/src/architecture/concrete/primitives/plane/PlaneCreator.ts
@@ -40,6 +40,10 @@ export class PlaneCreator extends BaseCreator {
     return this._domainObject;
   }
 
+  public override get canStartOnHover(): boolean {
+    return true;
+  }
+
   public override get preferIntersection(): boolean {
     return true;
   }
